+++ Abdm.Calculation/Models/Surface.cs
@@ -5,41 +5,41 @@ namespace Abdm.Calculation.Models
     public class Surface
     {
         /// <summary>
-        /// Массив точек, из которых состоит данная поверхность
+        /// Массив точек, из которых состоит поверхность влияния
         /// </summary>
         required public Vector3[] SurfacePoints { get; set; }
 
-        public float MinZ { get; set; }
+        public double MinZ { get; set; }
 
         /// <summary>
         /// Данные по опоре. Если чекпоинт не являтся опорой - массив пустой
         /// </summary>
-        public float[] PillarData { get; set; }
+        public double[] PillarData { get; set; }
 
         /// <summary>
         /// Максимальное значение всех точек по X
         /// </summary>
-        public float MaxX { get; set; }
+        public double MaxX { get; set; }
 
         /// <summary>
         /// Минимальное значение всех точек по X
         /// </summary>
-        public float MinX { get; set; }
+        public double MinX { get; set; }
 
         /// <summary>
         /// Максимальное значение всех точек по Y
         /// </summary>
-        public float MaxY { get; set; }
+        public double MaxY { get; set; }
 
         /// <summary>
         /// Минимальное значение всех точек по Y
         /// </summary>
-        public float MinY { get; set; }
+        public double MinY { get; set; }
 
         /// <summary>
         /// Максимальное значение всех точек по Z
         /// </summary>
-        public float MaxZ { get; set; }
+        public double MaxZ { get; set; }
 
         /// <summary>
         /// Перечисление, указывающее на то, как поверхность будет подвергаться нагрузке
@@ -50,26 +50,26 @@ namespace Abdm.Calculation.Models
         /// <summary>
         /// Проектная устойчивость структуры. Без учёта собственного веса
         /// </summary>
-        public float MyStrength { get; set; }
+        public double MyStrength { get; set; }
 
         /// <summary>
         /// Фиксированная нагрузка от собственного веса
         /// </summary>
-        public float СonstLoad { get; set; }
+        public double СonstLoad { get; set; }
 
         /// <summary>
         /// Нагрузка от пешеходов
         /// </summary>
-        public float PedestrianLoad { get; set; } 
+        public double PedestrianLoad { get; set; } 
 
         /// <summary>
         /// Другая нагрузка
         /// </summary>
-        public float OtherLoad { get; set; }
+        public double OtherLoad { get; set; }
 
         /// <summary>
         /// Коэффициент устойчивости. По дефолту всегда 1.
         /// </summary>
-        public float KStrength { get; set; }
+        public double KStrength { get; set; }
     }
 }
